requirements.txt
openai-whisper
sounddevice
soundfile
numpy
scipy
keyboard
pyperclip
pyautogui
pystray
pillow
plyer
